requirements.txt
alipay==0.7.4
amqp==5.0.6
ansicon==1.89.0
asgiref==3.3.4
asn1crypto==0.24.0
backports.csv==1.0.7
billiard==3.6.4.0
blessed==1.18.0
celery==5.0.5
certifi==2020.6.20
cffi==1.14.5
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
coreapi==2.3.3
coreschema==0.0.4
crypto==1.4.1
cryptography==3.4.6
defusedxml==0.5.0
demjson==2.2.4
diff-match-patch==20181111
Django==3.2
django-cors-headers==3.2.0
django-countries==5.3.3
django-crispy-forms==1.7.2
django-debug-toolbar==1.10.1
django-filter==2.4.0
django-formtools==2.1
django-guardian==2.0.0
django-import-export==1.2.0
django-jet==1.0.8
django-qiniu-storage==2.3.1
django-ranged-response==0.2.0
django-redis==4.12.1
django-reversion==3.0.4
django-simple-captcha==0.5.12
django-simpleui==2021.4.1
djangorestframework==3.12.1
djangorestframework-jwt==1.11.0
DjangoUeditor==1.8.143
dnspython==1.16.0
et-xmlfile==1.0.1
eventlet==0.30.2
future==0.17.1
gpustat==1.0.0.dev1
greenlet==1.0.0
httplib2==0.13.0
idna==2.10
itypes==1.2.0
jdcal==1.4.1
Jinja2==2.11.2
jinxed==1.0.1
kombu==5.0.2
Markdown==3.1.1
MarkupSafe==1.1.1
msgpack==1.0.2
mysqlclient==2.0.1
Naked==0.1.31
numpy==1.19.1
nvidia-ml-py3==7.352.0
odfpy==1.4.0
openpyxl==2.6.2
Pillow==7.2.0
prompt-toolkit==3.0.18
psutil==5.8.0
pyasn1==0.4.8
pycparser==2.20
pycryptodome==3.10.1
pycryptodomex==3.9.4
pyinstall==0.1.4
PyJWT==1.7.1
pyOpenSSL==19.1.0
python-alipay-sdk==3.0.1
pytz==2020.1
PyYAML==5.1.1
qcloudsms-py==0.1.4
qiniu==7.2.4
redis==3.5.3
requests==2.24.0
rsa==4.7.2
shellescape==3.8.1
simplepro==2.3.4
six==1.15.0
sqlparse==0.3.1
tablib==0.13.0
tencentcloud-sdk-python==3.0.365
tyadmin-api-cli==0.5.1
uritemplate==3.0.1
urllib3==1.25.10
vine==5.0.0
wcwidth==0.2.5
xlrd==1.2.0
xlwt==1.3.0
you-get==0.4.1432
